requirements.txt
torch>=2.2
pandas>=2.1
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
shap>=0.44
matplotlib>=3.8
rdkit-pypi>=2022.9.5
